requirements.txt
# -----------------------------
# Core config / typing
# -----------------------------
python-dotenv>=1.0,<2.0
pydantic>=2.4,<3.0

# -----------------------------
# HTTP + resilience
# -----------------------------
requests>=2.31,<3.0
tenacity>=8.2,<9.0

# -----------------------------
# Binance live ingestion
# -----------------------------
websockets>=12.0,<14.0

# -----------------------------
# PostgreSQL
# -----------------------------
psycopg[binary]>=3.1,<4.0

# -----------------------------
# Data handling / features
# -----------------------------
pandas>=2.0,<3.0
numpy>=1.26,<3.0

# -----------------------------
# Monitoring / dashboard
# -----------------------------
streamlit>=1.32,<2.0
streamlit-plotly-events>=0.0.5
plotly>=5.18,<6.0

# -----------------------------
# Logging helpers (optional)
# -----------------------------
python-json-logger>=2.0,<3.0